+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 1.1 (Unreleased)
 ================
 
+API Breaking Changes
+--------------------
+- `~ndcube.NDCubeBase.crop_by_extra_coord` API has been broken and
+  replaced.
+  Old version:
+  ``crop_by_extra_coord(min_coord_value, interval_width, coord_name)``.
+  New version:
+  ``crop_by_extra_coord(coord_name, min_coord_value,  max_coord_value)``.
+  [#140]
+
 New Features
 ------------
 - Created a new `~ndcube.NDCubeBase` which has all the functionality